2. ПРИКЛАДНАЯ РАЗРАБОТКА НА PYTHON/2_09_FASTAPI/requirements.txt
fastapi
uvicorn
loguru
pydentic
sqlalchemy
psycopg2-binary